code refactor
I never intended to make this a large extension, originally I only planned to add the case page odds and went with a quick and lazy solution for it.

I have since decided to make it a fully fledged extension with lots of functionality, and I also plan to make it open source on GitHub.

Therefore I have refactored the code to make it easier to edit and expand, and also so people are less likely to make fun of my low quality code.